opcodes/asmjs-dis.c: create
Disassembly is somewhat tricky because our object file format doesn't
contain JavaScript code (which is what we want the disassembler to emit)
directly in the .text section: instead, the .text section contais
pointers (padded to 8 bytes for future 64-bit-ness) to the beginning and
end of the actual JavaScript code.

In particular, this means that GDB must be told to trust read-only
sections to disassemble code correctly.